LibCli.Adapters

Description

Documentation

data KbEntry

KbEntry is an entry on the file containing the Knowledge Base.

Constructors

KbEntry

Fields

abbreviation :: String

expansion :: String

mapEntries :: KbEntry -> (Keyword, Keyword)

mapEntries maps entries to a pair of (abbreviation, expansion)

mapKeywordPair :: (Keyword, Keyword) -> KbEntry

mapKeywords maps keyword pairs to their entries

getKnowledgeBase

Arguments

:: Vector KbEntry
it is the result of the low level parsing of the file, represented as a vector of entries

-> KnowledgeBaseStructure

getKnowledgeBase parses from a CSV file to a KnowledgeBaseStructure.