LibCli.HandlerUtils

Description

Documentation

loadKb :: FilePath -> IO (Either Error KnowledgeBaseStructure)

Function to load the Knowledge Base from the specified file. Supports only CSV files.

loadInput :: FilePath -> IO (Either Error String)

Function to load the abbreviation input file. Can be any file with contents loadable as String.

doExpansion :: KnowledgeBaseStructure -> String -> IO String

Performs the expansion logic on the provided string.

formatRecord :: (Keyword, Keyword) -> String

Pretty prints the Keyword pair (key, value) from the Knowledge Base

dump :: FilePath -> KnowledgeBaseStructure -> IO ()

Dumps the Knowledge Base to the specified file path. Writes the KB out in CSV format.

returnOutput :: FilePath -> String -> IO ()

Given a FilePath and a string, write the string to the FilePath. Throws an error if no valid path is specified. Write the given string to the file path otherwise.

getInOutFilePaths

Arguments

:: Maybe FilePath
Input file path

-> Maybe FilePath
Output file path

-> Bool
Whether working in in-place mode.

-> Either Error (FilePath, FilePath)
Either error or the correct (Input, Output) pair.

Gets the pair of paths to use depending on the flag. Returns an error if a pair of (Input, Output) file paths is not possible. Otherwise, returns the pair of file paths in the following order: (Input, Output).

getKnowledgeBaseFilePath :: Maybe FilePath -> Either Error FilePath

Handles the retrieval of Knowledge Base file paths. If nothing is provided, raises an error. Otherwise, returns the given file path.

unwrapIOError :: Either Error a -> IO a

Returns the value from a potentially failing computation.

Abort with the error message if it was an error.

The text message is added to the Error as additional context before aborting.

Panics: if Error

Example:

>>> unwrapIOError $ Left (StandardError "oh no!")
*** Exception: StandardError "oh no!"

>>> unwrapIOError $ Right 42
42