LibCli.Handlers

Description

Documentation

Handlers are the core part of the CLI, as they handle the commands and how they should behave.

Expansion Operations

Here we adopt the following convention:

fp is a FilePath

m stands for Maybe, so mfp is `Maybe FilePath`

What precedes is the following:

kb is the KnowledgeBaseStructure and if followed by _fp it means that the underlying contents of the File at FilePath fp is a kb

in is the same, input file

o is output file

TODO io is input/output file (used in inplace mode)

expandHandler

Arguments

:: Maybe FilePath
KB file path

-> String
Abbreviation to expand

-> IO ()
Writes the expansion result to the STDOUT.

Expand command handler. Deal with single input expansion without reading an input file.

replaceHandler

Arguments

:: Maybe FilePath
KB file path

-> Maybe FilePath
Input file path

-> Maybe FilePath
Output file path

-> Bool
Whether inplace mode is used or not

-> IO ()
Writes the modified file out to the specified location

Replace command handler. Does all the replacing heavy-lifting. Loads the required files, produces the expansions and writes the output file.

Knowledge Base CRUD

addHandler

Arguments

:: Maybe FilePath
KB file path

-> String
Abbreviation keyword

-> String
Expansion keyword

-> IO ()
Writes the full contents of the KB to the STDOUT.

Add command handler. Adds new abbreviation with the associated expansion to the KB.

updateHandler

Arguments

:: Maybe FilePath
KB file path

-> String
Abbreviation keyword

-> String
Expansion keyword

-> IO ()
Writes the full contents of the KB to the STDOUT.

Update command handler. Updates an existing abbreviation in the KB.

deleteHandler

Arguments

:: Maybe FilePath
KB file path

-> String
Abbreviation keyword

-> IO ()
Writes the full contents of the KB to the STDOUT.

Delete command handler. Deletes an existing abbreviation from the KB.

listHandler

Arguments

:: Maybe FilePath
KB file path

-> IO ()
Writes the full contents of the KB to the STDOUT.

List command handler. Displays all the contents of the specified Knowledge base.