LibCore.KnowledgeBase

Description

Documentation

type KnowledgeBaseStructure = Map Keyword Keyword

build :: [(Keyword, Keyword)] -> KnowledgeBaseStructure

Build a new Knowledge Base from the supplied keyword pair list.

listAll :: KnowledgeBaseStructure -> [(Keyword, Keyword)]

Get the list of all the stored records in the Knowledge Base.

Examples:

>>> testKB = build [(Keyword {keyword = "brb", plural = False}, Keyword {keyword = "be right back", plural = False})]
>>> listAll testKB
[(Keyword {keyword = "brb", plural = False},Keyword {keyword = "be right back", plural = False})]

get :: KnowledgeBaseStructure -> Keyword -> Either Error Keyword

Retrieve a single element by its keyword.

Examples

>>> testKB = build [(Keyword {keyword = "brb", plural = False}, Keyword {keyword = "be right back", plural = False})]
>>> get testKB (Keyword {keyword = "brb", plural = False})
Right (Keyword {keyword = "be right back", plural = False})

>>> get testKB (Keyword {keyword = "beb", plural = False})
Left (StandardError "no record found for this keyword : Keyword {keyword = \"beb\", plural = False}")

add :: KnowledgeBaseStructure -> Keyword -> Keyword -> (Keyword, KnowledgeBaseStructure)

Add new item to the Knowledge Base. Adds the item to the KB as a new record if the key was notseen before, and replaces the existing record otherwise.

Examples

>>> testKB = build [(Keyword {keyword = "brb", plural = False}, Keyword {keyword = "be right back", plural = False})]
>>> add testKB (Keyword {keyword = "btw", plural = False}) (Keyword {keyword = "by the way", plural = False})
(Keyword {keyword = "by the way", plural = False},fromList [(Keyword {keyword = "brb", plural = False},Keyword {keyword = "be right back", plural = False}),(Keyword {keyword = "btw", plural = False},Keyword {keyword = "by the way", plural = False})])

>>> add testKB (Keyword {keyword = "btw", plural = False}) (Keyword {keyword = "by the way!", plural = False})
(Keyword {keyword = "by the way!", plural = False},fromList [(Keyword {keyword = "brb", plural = False},Keyword {keyword = "be right back", plural = False}),(Keyword {keyword = "btw", plural = False},Keyword {keyword = "by the way!", plural = False})])

put :: KnowledgeBaseStructure -> Keyword -> Keyword -> Either Error (Keyword, KnowledgeBaseStructure)

Updates the record at given key with the new value. Only full updates are supported. If the key is not in the KB, returns an error.

Examples:

>>> testKB = build [(Keyword {keyword = "brb", plural = False}, Keyword {keyword = "be right back", plural = False})]
>>> put testKB (Keyword {keyword = "brb", plural = False}) (Keyword {keyword = "BE RIGHT BACK", plural = False})
Right (Keyword {keyword = "BE RIGHT BACK", plural = False},fromList [(Keyword {keyword = "brb", plural = False},Keyword {keyword = "BE RIGHT BACK", plural = False})])

>>> put testKB (Keyword {keyword = "btw", plural = False}) (Keyword {keyword = "by the way!", plural = False})
Left (StandardError "no record found for this keyword : Keyword {keyword = \"btw\", plural = False}")

remove :: KnowledgeBaseStructure -> Keyword -> Either Error KnowledgeBaseStructure

Removes an existing item at the given key form the Knowledge Base. If the key is not present, return as error.

Examples:

>>> testKB = build [(Keyword {keyword = "brb", plural = False}, Keyword {keyword = "be right back", plural = False})]
>>> remove testKB (Keyword {keyword = "brb", plural = False})
Right (fromList [])

>>> remove testKB (Keyword {keyword = "btw", plural = False})
Left (StandardError "no record found for this keyword : Keyword {keyword = \"btw\", plural = False}")