LibCore.Models

Description

Documentation

data Token

We parse a string into a NoToken if it does not match the syntax for | parsing. The DoMap is a string that has to be mapped. In our proposals, | the DoMap strings start with @@

Constructors

NoToken String

DoMap Keyword

data AKeyword a

Abstract and parameterised Keyword datatype. Implements Functor and Applicative typeclasses.

Examples:

>>> pure "hello" :: AKeyword String
Keyword {keyword = "hello", plural = False}

>>> (map Data.Char.toUpper) <$> (pure "hello" :: AKeyword String)
Keyword {keyword = "HELLO", plural = False}

>>> (pure (map Data.Char.toLower)) <*> (pure "HELLO" :: AKeyword String)
Keyword {keyword = "hello", plural = False}

Constructors

Keyword

Fields

keyword :: a

plural :: Bool

type Keyword = AKeyword String

A Keyword is a Token that can be mapped. A Keyword can be plural

newtype Error

Reserved domain-level error indications.

Constructors

StandardError String