LibCore.Parser

Description

Documentation

type ParseStructure = [Token]

abbSymbol :: String

The abbreviation symbol. Can later be made configurable

pluralSymbol :: String

The plural symbol. String between the abbSymbol and the pluralSymbol are interpreted as plurals

whitespace :: Parser ()

Consume whitespace characters

lexeme :: Parser a -> Parser a

Lexeme parsers consume the whitespace after them

doParse :: String -> ParseStructure

Given a string, parse it. This function can throw an error if parsing fails

parseInput :: String -> Either ParseError ParseStructure

Map a string to a list of Tokens. For example: >>> parseInput "@bob" Right [DoMap (Keyword {keyword = "bob", plural = False})] >>> parseInput "@fw's" Right [DoMap (Keyword {keyword = "fw", plural = True})] >>> parseInput "hello!" Right [NoToken "hello!"]

mainParser :: String -> String -> Parser ParseStructure

The main parser tries to consume all input into a ParseStructure, given an | abbreviation symbol and a plural symbol

notSpace :: Char -> Bool

Inverse of the isSpace function from Data.Char

pluralAbbrParser :: String -> String -> Parser Token

Given an abbreviation string s and a plural end string p, parse the string between it

abbrParser :: String -> Parser Token

Given an abbreviation string s, parse the string after it

noAbbrParser :: Parser Token

Parse any string into a token

punctuationParser :: Parser Token

Parser to process punctuation. We added this because abbreviations followed by | punctuation without a space, such as 'hello!' were parsed as NoToken