LibCli.Handlers

Description

Documentation

loadKb :: FilePath -> IO (Either Error KnowledgeBaseStructure)

loadInput :: FilePath -> IO (Either Error String)

doExpansion :: KnowledgeBaseStructure -> String -> IO String

formatRecord :: (Keyword, Keyword) -> String

dump :: FilePath -> KnowledgeBaseStructure -> IO ()

expandHandler

Arguments

:: Maybe FilePath
KB file path

-> String
Abbreviation to expand

-> IO ()
Writes the expansion result to the STDOUT.

Expand command handler. Deal with single input expansion without reading an input file.

replaceHandler

Arguments

:: Maybe FilePath
KB file path

-> Maybe FilePath
Input file path

-> Maybe FilePath
Output file path

-> IO ()
Writes the modified file out to the specified location

Replace command handler. Does all the replacing heavy-lifting. Loads the required files, produces the expansions and writes the output file.

addHandler

Arguments

:: Maybe FilePath
KB file path

-> String
Abbreviation keyword

-> String
Expansion keyword

-> IO ()
Writes the full contents of the KB to the STDOUT.

Add command handler. Adds new abbreviation with the associated expansion to the KB.

updateHandler

Arguments

:: Maybe FilePath
KB file path

-> String
Abbreviation keyword

-> String
Expansion keyword

-> IO ()
Writes the full contents of the KB to the STDOUT.

Update command handler. Updates an existing abbreviation in the KB.

deleteHandler

Arguments

:: Maybe FilePath
KB file path

-> String
Abbreviation keyword

-> IO ()
Writes the full contents of the KB to the STDOUT.

Delete command handler. Deletes an existing abbreviation from the KB.

listHandler

Arguments

:: Maybe FilePath
KB file path

-> IO ()
Writes the full contents of the KB to the STDOUT.

List command handler. Displays all the contents of the specified Knowledge base.