LibCore.KnowledgeBase

Description

Documentation

type KnowledgeBaseStructure = Map Keyword Keyword

getKnowledgeBase :: KnowledgeBaseStructure

buildKnowledgeBase :: [(Keyword, Keyword)] -> KnowledgeBaseStructure

listAll :: KnowledgeBaseStructure -> [(Keyword, Keyword)]

Get the list of all the stored records in the Knowledge Base.

Examples:

>>> testKB = buildKnowledgeBase [(Keyword {keyword = "brb", plural = False}, Keyword {keyword = "be right back", plural = False})]
>>> listAll testKB
[(Keyword {keyword = "brb", plural = False},Keyword {keyword = "be right back", plural = False})]

get :: KnowledgeBaseStructure -> Keyword -> Either Error Keyword

Retrieve a single element by its keyword.

Examples

>>> testKB = buildKnowledgeBase [(Keyword {keyword = "brb", plural = False}, Keyword {keyword = "be right back", plural = False})]
>>> get testKB (Keyword {keyword = "brb", plural = False})
Right (Keyword {keyword = "be right back", plural = False})

>>> get testKB (Keyword {keyword = "beb", plural = False})
Left (StandardError "no record found for this keyword")

add :: KnowledgeBaseStructure -> Keyword -> Keyword -> (Keyword, KnowledgeBaseStructure)

Add new item to the Knowledge Base. Adds the item to the KB as a new record if the key was notseen before, and replaces the existing record otherwise.

Examples

>>> testKB = buildKnowledgeBase [(Keyword {keyword = "brb", plural = False}, Keyword {keyword = "be right back", plural = False})]
>>> add testKB (Keyword {keyword = "btw", plural = False}) (Keyword {keyword = "by the way", plural = False})
(Keyword {keyword = "by the way", plural = False},fromList [(Keyword {keyword = "brb", plural = False},Keyword {keyword = "be right back", plural = False}),(Keyword {keyword = "btw", plural = False},Keyword {keyword = "by the way", plural = False})])

>>> add testKB (Keyword {keyword = "btw", plural = False}) (Keyword {keyword = "by the way!", plural = False})
(Keyword {keyword = "by the way!", plural = False},fromList [(Keyword {keyword = "brb", plural = False},Keyword {keyword = "be right back", plural = False}),(Keyword {keyword = "btw", plural = False},Keyword {keyword = "by the way!", plural = False})])

update :: KnowledgeBaseStructure -> Keyword -> Either Error Keyword

delete :: KnowledgeBaseStructure -> Keyword -> Either Error Bool