LibCore.Models

Description

Documentation

data Token

We parse a string into a NoToken if it does not match the syntax for | parsing. The DoMap is a string that has to be mapped. In our proposals, | the DoMap strings start with @@

Constructors

NoToken String

DoMap Keyword

data Keyword

Constructors

Keyword

Fields

keyword :: String

plural :: Bool